Replace implementations of `ICommandHandler` by moving the logic to the commands
This is the first step of no longer relying on `System.CommandLine.Hosting`.
