+++ src/FaluCli/Commands/Messages/MessagesSendCommand.cs
@@ -1,10 +1,14 @@
-﻿using Res = Falu.Properties.Resources;
+﻿using Falu.Client;
+using Falu.MessageBatches;
+using Falu.Messages;
+using Falu.MessageTemplates;
+using Res = Falu.Properties.Resources;
 
 namespace Falu.Commands.Messages;
 
-public abstract class AsbtractMessagesSendCommand : Command
+public abstract class AbstractMessagesSendCommand : Command
 {
-    public AsbtractMessagesSendCommand(string name, string? description = null) : base(name, description)
+    protected AbstractMessagesSendCommand(string name, string? description = null) : base(name, description)
     {
         this.AddOption<string[]>(["--to", "-t"],
                                  description: "Phone number(s) you are sending to, in E.164 format.",
@@ -50,12 +54,14 @@ public abstract class AsbtractMessagesSendCommand : Command
         this.AddOption(["--schedule-delay", "--delay"],
                        description: "The delay (in ISO8601 duration format) to be applied by the server before sending the message(s).\r\nExample: PT10M for 10 minutes",
                        format: Constants.Iso8061DurationFormat);
+
+        this.SetHandler(HandleAsync);
     }
 
     private static string? ValidateNumbers(string optionName, string[] numbers)
     {
         // ensure not more than 1k messages in a batch
-        var limit = 1_000;
+        const int limit = 1_000;
         if (numbers.Length > limit)
         {
             return string.Format(Res.TooManyMessagesToBeSent, limit);
@@ -72,9 +78,157 @@ public abstract class AsbtractMessagesSendCommand : Command
 
         return null;
     }
+
+    private static async Task HandleAsync(InvocationContext context)
+    {
+        var cancellationToken = context.GetCancellationToken();
+        var command = context.ParseResult.CommandResult.Command;
+        var client = context.GetRequiredService<FaluCliClient>();
+        var logger = context.GetRequiredService<ILoggerFactory>().CreateLogger(command.GetType());
+
+        // ensure both to and file are not null or empty
+        var tos = context.ParseResult.ValueForOption<string[]>("--to");
+        var filePath = context.ParseResult.ValueForOption<string>("--file");
+        if ((tos is null || tos.Length == 0) && string.IsNullOrWhiteSpace(filePath))
+        {
+            logger.LogError("A CSV file path must be specified or the destinations using the --to option.");
+            context.ExitCode = -1;
+            return;
+        }
+
+        // ensure both to and file are not specified
+        if (tos is not null && tos.Length > 0 && !string.IsNullOrWhiteSpace(filePath))
+        {
+            logger.LogError("Either specify the CSV file path or destinations not both.");
+            context.ExitCode = -1;
+            return;
+        }
+
+        // read the numbers from the CSV file
+        if (tos is null || tos.Length == 0)
+        {
+            tos = File.ReadAllText(filePath!)
+                      .Replace("\r\n", ",")
+                      .Replace("\r", ",")
+                      .Replace("\n", ",")
+                      .Split(',', StringSplitOptions.RemoveEmptyEntries);
+        }
+
+        var stream = context.ParseResult.ValueForOption<string>("--stream")!;
+
+        // ensure both media URL and medial file Id are not specified
+        var mediaUrl = context.ParseResult.ValueForOption<Uri?>("--media-url");
+        var mediaFileId = context.ParseResult.ValueForOption<string?>("--media-file-id");
+        if (mediaUrl is not null && mediaFileId is not null)
+        {
+            logger.LogError("Media URL and File ID cannot be specified together.");
+            context.ExitCode = -1;
+            return;
+        }
+
+        var media = mediaUrl is not null || mediaFileId is not null
+                  ? new[] { new MessageCreateRequestMedia { Url = mediaUrl?.ToString(), File = mediaFileId, }, }
+                  : null;
+
+        // ensure both time and delay are not specified
+        var time = context.ParseResult.ValueForOption<DateTimeOffset?>("--schedule-time");
+        var delay = context.ParseResult.ValueForOption<string?>("--schedule-delay");
+        if (time is not null && delay is not null)
+        {
+            logger.LogError("Schedule time and delay cannot be specified together.");
+            context.ExitCode = -1;
+            return;
+        }
+
+        // make the schedule
+        var schedule = time is not null
+                     ? (MessageCreateRequestSchedule)time
+                     : delay is not null
+                        ? (MessageCreateRequestSchedule)delay
+                        : null;
+
+        string? body = null;
+        MessageCreateRequestTemplate? template = null;
+        if (command is MessagesSendRawCommand)
+        {
+            body = context.ParseResult.ValueForOption<string>("--body")!; // marked required in the command
+        }
+        else if (command is MessagesSendTemplatedCommand)
+        {
+            var id = context.ParseResult.ValueForOption<string>("--id");
+            var alias = context.ParseResult.ValueForOption<string>("--alias");
+            var language = context.ParseResult.ValueForOption<string>("--language");
+            var modelJson = context.ParseResult.ValueForOption<string>("--model")!; // marked required in the command
+            var model = new MessageTemplateModel(System.Text.Json.Nodes.JsonNode.Parse(modelJson)!.AsObject());
+
+            // ensure both id and alias are not null
+            if (string.IsNullOrWhiteSpace(id) && string.IsNullOrWhiteSpace(alias))
+            {
+                logger.LogError("A template identifier or template alias must be provided when sending a templated message.");
+                context.ExitCode = -1;
+                return;
+            }
+
+            // ensure both id and alias are not specified
+            if (!string.IsNullOrWhiteSpace(id) && !string.IsNullOrWhiteSpace(alias))
+            {
+                logger.LogError("Either specify the template identifier or template alias not both.");
+                context.ExitCode = -1;
+                return;
+            }
+
+            template = new MessageCreateRequestTemplate { Id = id, Alias = alias, Language = language, Model = model, };
+        }
+        else throw new InvalidOperationException($"Command of type '{command.GetType().FullName}' is not supported here.");
+
+        // if there is only a single number, send a single message, otherwise use the batch
+        if (tos.Length == 1)
+        {
+            var target = tos[0];
+            var request = new MessageCreateRequest
+            {
+                To = target,
+                Body = body,
+                Template = template,
+                Stream = stream,
+                Media = media,
+                Schedule = schedule,
+            };
+            var rr = await client.Messages.CreateAsync(request, cancellationToken: cancellationToken);
+            rr.EnsureSuccess();
+
+            var response = rr.Resource!;
+            logger.LogInformation("Scheduled {MessageId} for sending at {Scheduled:f}.", response.Id, (response.Schedule?.Time ?? response.Created).ToLocalTime());
+        }
+        else
+        {
+            var request = new MessageBatchCreateRequest
+            {
+                Messages =
+                [
+                    new MessageBatchCreateRequestMessage
+                    {
+                        Tos = tos,
+                        Body = body,
+                        Template = template,
+                        Media = media,
+                    },
+                ],
+                Stream = stream,
+                Schedule = schedule,
+            };
+            var rr = await client.MessageBatches.CreateAsync(request, cancellationToken: cancellationToken);
+            rr.EnsureSuccess();
+
+            var response = rr.Resource!;
+            var ids = response.Messages!;
+            logger.LogInformation("Scheduled {Count} messages for sending at {Scheduled:f}.", ids.Count, (response.Schedule?.Time ?? response.Created).ToLocalTime());
+            logger.LogDebug("Message Id(s):\r\n- {Ids}", string.Join("\r\n- ", ids));
+        }
+    }
 }
 
-public class MessagesSendRawCommand : AsbtractMessagesSendCommand
+public class MessagesSendRawCommand : AbstractMessagesSendCommand
 {
     public MessagesSendRawCommand() : base("raw", "Send a message with the body defined.")
     {
@@ -84,7 +238,7 @@ public class MessagesSendRawCommand : AsbtractMessagesSendCommand
     }
 }
 
-public class MessagesSendTemplatedCommand : AsbtractMessagesSendCommand
+public class MessagesSendTemplatedCommand : AbstractMessagesSendCommand
 {
     public MessagesSendTemplatedCommand() : base("templated", "Send a templated message.")
     {
