+++ src/FaluCli/Extensions/IHostBuilderExtensions.cs
@@ -4,7 +4,6 @@ using Falu.Config;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
 using System.CommandLine.Hosting;
-using System.Diagnostics.CodeAnalysis;
 
 namespace Microsoft.Extensions.Hosting;
 
@@ -51,16 +50,5 @@ internal static class IHostBuilderExtensions
         });
 
         return builder;
-    }
-
-    // this exists to work around trimming which does not keep constructors by default in the System.CommandLine.Hosting library
-    // though Microsoft.Extensions.DependencyInjection does
-    public static IHostBuilder UseCommandHandlerTrimmable<
-        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] TCommand,
-        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)] THandler>(this IHostBuilder builder)
-        where TCommand : Command
-        where THandler : ICommandHandler
-    {
-        return builder.UseCommandHandler<TCommand, THandler>();
     }
 }
